Move manage leave page to employee management
- Moved ManageLeaveRequestsPage.jsx from client/src/pages/hr/ to client/src/pages/employees/
- Updated route in App.jsx to reflect the new file location and path (/employees/manage-leave-requests).
- Updated Sidebar.jsx to move 'Manage Leave Requests' under 'Employee Management' and use the new path.
- Removed dynamic sidebar item generation for 'Manage Leave Requests' from MainLayout.jsx as it's now static in Sidebar.jsx.
